# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MySchool API",
    "version": "2.2.0",
    "description": "API REST complète pour MySchool e-learning. \n        Gestion des utilisateurs, cours, chapitres, leçons, exercices, inscriptions et instructeurs.\n        Support complet des requêtes filtrées et index Firestore optimisés.",
    "contact": {
      "name": "MySchool Support",
      "email": "support@myschool.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://us-central1-myschool-f862b.cloudfunctions.net",
      "description": "Serveur de production (Firebase Functions)"
    },
    {
      "url": "http://localhost:3000",
      "description": "Serveur de développement local"
    }
  ],
  "components": {
    "schemas": {
...
      "Course": {
        "type": "object",
        "required": [
          "title",
          "description",
          "level",
          "niveauScolaire",
          "classe"
        ],
        "properties": {
          "title": {
            "type": "string",
            "example": "Introduction à JavaScript"
          },
          "description": {
            "type": "string",
            "example": "Apprendre les bases de JavaScript"
          },
          "image": {
            "type": "string",
            "description": "URL de l'image du cours"
          },
          "level": {
            "type": "string",
            "enum": [
              "DEBUTANT",
              "INTERMEDIAIRE",
              "AVANCE"
            ],
            "example": "DEBUTANT"
          },
          "niveauScolaire": {
            "type": "string",
            "enum": [
              "ELEMENTAIRE",
              "MOYEN",
              "SECONDAIRE",
              "UNIVERSITAIRE"
            ],
            "example": "MOYEN"
          },
          "classe": {
            "type": "string",
            "example": "3ème (BFEM)",
            "description": "Classe cible du cours (obligatoire)"
          },
          "matiereId": {
            "type": "string",
            "description": "ID de la matière associée"
          },
          "duration": {
            "type": "number",
            "example": 120,
            "description": "Durée en minutes"
          },
          "price": {
            "type": "number",
            "example": 0,
            "description": "Prix (0 = gratuit)"
          },
          "category": {
            "type": "string",
            "example": "Programmation"
          },
          "type": {
            "type": "string",
            "enum": [
              "VIDEO",
              "En ligne",
              "Hybride"
            ],
            "example": "En ligne"
          },
          "rating": {
            "type": "number",
            "example": 4.5
          },
          "documents": {
            "type": "array",
            "description": "Documents PDF associés au cours",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Cours_Chapitre1.pdf"
                },
                "url": {
                  "type": "string",
                  "example": "https://storage.googleapis.com/.../document.pdf"
                },
                "size": {
                  "type": "number",
                  "example": 2048576,
                  "description": "Taille en octets"
                },
                "uploadedAt": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Instructor": {
        "type": "object",
        "required": [
          "userId",
          "bio",
          "specialization"
        ],
        "properties": {
          "userId": {
            "type": "string"
          },
          "bio": {
            "type": "string",
            "example": "Développeur web avec 10 ans d'expérience"
          },
          "specialization": {
            "type": "string",
            "example": "web-development"
          },
          "rating": {
            "type": "number",
            "example": 4.5,
            "minimum": 0,
            "maximum": 5
          },
          "totalStudents": {
            "type": "number",
            "example": 150
          },
          "totalCourses": {
            "type": "number",
            "example": 5
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/courses/instructor/{instructorId}": {
      "get": {
        "tags": [
          "Courses"
        ],
        "summary": "Récupérer les cours d'un professeur",
        "parameters": [
          {
            "in": "path",
            "name": "instructorId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID du professeur"
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des cours du professeur",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Course"
                      }
                    },
                    "count": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/instructors": {
      "post": {
        "tags": [
          "Instructors"
        ],
        "summary": "Créer un nouvel instructeur",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Instructor"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Instructeur créé"
          }
        }
      },
      "get": {
        "tags": [
          "Instructors"
        ],
        "summary": "Récupérer tous les instructeurs",
        "responses": {
          "200": {
            "description": "Liste des instructeurs"
          }
        }
      }
    },
    "/api/instructors/{id}": {
      "get": {
        "tags": [
          "Instructors"
        ],
        "summary": "Récupérer un instructeur par ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Instructeur trouvé"
          }
        }
      },
      "put": {
        "tags": [
          "Instructors"
        ],
        "summary": "Mettre à jour un instructeur",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Instructor"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Instructeur mis à jour"
          }
        }
      },
      "delete": {
        "tags": [
          "Instructors"
        ],
        "summary": "Supprimer un instructeur",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Instructeur supprimé"
          }
        }
      }
    },
    "/api/instructors/course/{courseId}": {
      "get": {
        "tags": [
          "Instructors"
        ],
        "summary": "Récupérer les instructeurs d'un cours",
        "parameters": [
          {
            "in": "path",
            "name": "courseId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des instructeurs"
          }
        }
      }
    },
    "/api/instructors/expertise/{expertiseId}": {
      "get": {
        "tags": [
          "Instructors"
        ],
        "summary": "Récupérer les instructeurs par expertise",
        "parameters": [
          {
            "in": "path",
            "name": "expertiseId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des instructeurs"
          }
        }
      }
    },
...
    "/api/payments": {
      "post": {
        "summary": "Crée un nouveau paiement",
        "tags": [
          "Payments"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId",
                  "enrollmentId",
                  "courseId",
                  "amount",
                  "paymentMethod"
                ],
                "properties": {
                  "userId": {
                    "type": "string"
                  },
                  "enrollmentId": {
                    "type": "string"
                  },
                  "courseId": {
                    "type": "string"
                  },
                  "amount": {
                    "type": "number"
                  },
                  "currency": {
                    "type": "string",
                    "default": "XOF"
                  },
                  "paymentMethod": {
                    "type": "string",
                    "enum": [
                      "orange-money",
                      "wave",
                      "free-money",
                      "card"
                    ]
                  },
                  "customerPhoneNumber": {
                    "type": "string",
                    "description": "Requis pour orange-money"
                  },
                  "customerFirstName": {
                    "type": "string"
                  },
                  "customerLastName": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Paiement créé avec succès"
          }
        }
      },
      "get": {
        "summary": "Récupère tous les paiements",
        "tags": [
          "Payments"
        ],
        "responses": {
          "200": {
            "description": "Liste des paiements"
          }
        }
      }
    },
    "/api/payments/{id}": {
      "get": {
        "summary": "Récupère un paiement par ID",
        "tags": [
          "Payments"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paiement trouvé"
          }
        }
      },
      "delete": {
        "summary": "Supprime un paiement",
        "tags": [
          "Payments"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paiement supprimé"
          }
        }
      }
    },
    "/api/payments/user/{userId}": {
      "get": {
        "summary": "Récupère tous les paiements d'un utilisateur",
        "tags": [
          "Payments"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des paiements"
          }
        }
      }
    },
    "/api/payments/enrollment/{enrollmentId}": {
      "get": {
        "summary": "Récupère tous les paiements d'une inscription",
        "tags": [
          "Payments"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "enrollmentId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des paiements"
          }
        }
      }
    },
    "/api/payments/status/{status}": {
      "get": {
        "summary": "Récupère tous les paiements par statut",
        "tags": [
          "Payments"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "status",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "SUCCESS",
                "FAILED",
                "CANCELLED",
                "EXPIRED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des paiements"
          }
        }
      }
    },
    "/api/payments/{id}/check": {
      "get": {
        "summary": "Vérifie le statut d'un paiement Orange Money",
        "tags": [
          "Payments"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Statut du paiement"
          }
        }
      }
    },
    "/api/payments/{id}/cancel": {
      "post": {
        "summary": "Annule un paiement en attente",
        "tags": [
          "Payments"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paiement annulé"
          }
        }
      }
    },
    "/api/payments/webhook/orange-money": {
      "post": {
        "summary": "Webhook pour les notifications Orange Money",
        "tags": [
          "Payments"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook traité"
          }
        }
      }
    },
...
  }
}